Some changes, shadows don't work yet

+++ src/DirectionalLight.cpp
@@ -3,7 +3,7 @@
 #include "DirectionalLight.h"
 #include "Shader.h"
 
-DirectionalLight::DirectionalLight(GLfloat _width, GLfloat _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity, const glm::vec3& _direction)
+DirectionalLight::DirectionalLight(GLuint _width, GLuint _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity, const glm::vec3& _direction)
         : Light(_width, _height, _colour, _aIntensity, _dIntensity), direction(_direction) {
     lightProj = glm::ortho(-20.0f, 20.0f, -20.0f, 20.0f, 0.1f, 100.0f);
 }
