+++ src/DirectionalLight.h
@@ -15,7 +15,7 @@ struct UniformDirectionalLight {
 
 class DirectionalLight : public Light {
     public:
-        DirectionalLight(GLfloat _width, GLfloat _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity, const glm::vec3& _direction);
+        DirectionalLight(GLuint _width, GLuint _height, const glm::vec3& _colour, GLfloat _aIntensity, GLfloat _dIntensity, const glm::vec3& _direction);
         ~DirectionalLight();
         void UseLight(GLfloat _ambientIntensityLocation, GLfloat _ambienColourLocation, GLfloat _diffuseIntensityLocation, GLfloat _directionLocation) const;
         static void GetUDirectionalLight(const Shader& _shader, UniformDirectionalLight& _uDirectionalLight);
